merging from master in several steps so I can handle
031798f specially

Conflicts:
	src/test/org/nlogo/headless/TestCompileBenchmarks.scala

+++ src/main/org/nlogo/lab/ProtocolSaver.scala
@@ -76,9 +76,9 @@ object ProtocolSaver
           elementWithAttributes(
             "steppedValueSet",
             attributes(("variable", valueSet.variableName),
-                       ("first", Dump.number(steppedValueSet.firstValue)),
-                       ("step", Dump.number(steppedValueSet.step)),
-                       ("last", Dump.number(steppedValueSet.lastValue))))
+                       ("first", Dump.number(steppedValueSet.firstValue.toDouble)),
+                       ("step", Dump.number(steppedValueSet.step.toDouble)),
+                       ("last", Dump.number(steppedValueSet.lastValue.toDouble))))
         case enumeratedValueSet: EnumeratedValueSet =>
           hd.startElement("", "", "enumeratedValueSet",
                           attributes(("variable", valueSet.variableName)))
